+++ importlib_metadata/docs/changelog.rst
@@ -4,6 +4,8 @@
 
 0.7 (unreleased)
 ================
+* Fixed issue where packages with dashes in their names would
+  not be discovered. Closes #21.
 
 0.6 (2018-10-07)
 ================